+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 
+  - Kubernetes connector - support deployment logs action
   - Fix regression stop docker application
 
 ## [2.3.9] - 2019-12-09